github-to-social-automation/requirements.txt
Flask==2.3.3
PyGitHub==1.59.1
requests==2.31.0
groq==0.9.0
python-dotenv==1.0.0
schedule==1.2.0
